fix: moving graphql from plugin to core
