+++ packages/cli/package.json
@@ -27,6 +27,7 @@
     "@webcomponents/webcomponentsjs": "^2.3.0",
     "apollo-boost": "^0.4.4",
     "apollo-cache-inmemory": "^1.6.3",
+    "apollo-server": "^2.9.9",
     "axios": "^0.19.0",
     "babel-loader": "^8.0.5",
     "chalk": "^2.4.2",
