+++ packages/cli/src/graphql/schemas.js
@@ -1,7 +1,8 @@
 const { gql } = require('apollo-server-express');
 
-const { menuTypeDefs, menuResolvers } = require('./schemas/menu-schema');
+const { menuTypeDefs, menuResolvers } = require('./menu-schema');
 
+// eventually it will loop through and merge plugin schemas
 exports.typeDefs = gql`
   ${menuTypeDefs}
 `;
