+++ packages/plugin-google-analytics/test/cases/option-anonymous/option-anonymous.spec.js
@@ -62,7 +62,7 @@ describe('Build Greenwood With: ', async function() {
       });
 
       it('should be one inline <script> tag', function() {
-        expect(inlineScript.length).to.be.equal(1);
+        expect(inlineScript.length).to.be.equal(2);
       });
 
       it('should have the expected code with users analyicsId', function() {
@@ -75,7 +75,7 @@ describe('Build Greenwood With: ', async function() {
             gtag('config', '${mockAnalyticsId}');
         `;
 
-        expect(inlineScript[0].textContent).to.contain(expectedContent);
+        expect(inlineScript[1].textContent).to.contain(expectedContent);
       });
     });
