--- packages/plugin-graphql/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@greenwood/plugin-graphql",
-  "version": "0.4.0",
-  "description": "A plugin for Greenwood to enable use of GraphQL",
-  "repository": "https://github.com/ProjectEvergreen/greenwood/tree/master/packages/plugin-graphql",
-  "author": "Grant Hutchinson <grant@hutchdev.ca>",
-  "license": "MIT",
-  "keywords": [
-    "Greenwood",
-    "Web Components",
-    "Lit Element",
-    "Lit Html",
-    "Static Site Generator",
-    "GraphQL"
-  ],
-  "main": "src/server.js",
-  "files": [
-    "src/"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "peerDependencies": {
-    "@greenwood/cli": "^0.4.0"
-  },
-  "devDependencies": {
-    "@greenwood/cli": "^0.4.0"
-  },
-  "dependencies": {
-    "apollo-server": "^2.9.7",
-    "graphql": "^14.5.8"
-  }
-}
